Compile the shell on the two platforms that met this code today
The matrix has named macOS and Windows since it existed, but the terminal and
the atomic save were born in a tree the matrix never saw, so today was their
first contact. Two findings. rustix's RawMode is u32 on Linux and u16 on
Apple and the BSDs, so the umask sampling widens it once where the mask
arithmetic starts. And rustix gates its fs module out on Windows entirely,
which takes the PTY with it: the pty module is unix-only now, and on Windows
the terminal toggle still opens the view -- it answers with the labelled
failure the ExecSession seam was built to carry, rather than losing the
action or pretending to a shell it cannot fork.

+++ crates/k10s-shell/src/fs.rs
@@ -433,7 +433,10 @@ fn current_umask() -> u32 {
     *UMASK.get_or_init(|| {
         let mode = rustix::process::umask(rustix::fs::Mode::empty());
         rustix::process::umask(mode);
-        mode.bits()
+        // `RawMode` is u32 on Linux and u16 on Apple and the BSDs; widen it
+        // here so the mask arithmetic stays one type on every platform.
+        #[allow(clippy::useless_conversion)]
+        u32::from(mode.bits())
     })
 }
